Move Tuples to CoreTests
